// src/shared/ComponentsUI/components/Button/enums.ts
export enum Color {
  primary = 'primary',
  secondary = 'secondary',
  success = 'success',
  warning = 'warning',
  error = 'error',
}

export enum Size {
  small = 'small',
  medium = 'medium',
  large = 'large',
}

export enum Variant {
  text = 'text',
  contained = 'contained',
  outlined = 'outlined',
}